next ok

+++ Asteroids.Standard/Components/Bullet.cs
@@ -56,8 +56,8 @@ namespace Asteroids.Standard.Components
             var sinVal = Math.Sin(Radians);
             var cosVal = Math.Cos(Radians);
 
-            VelocityX = (int)(InitialVelocity * sinVal) + parentShip.GetVelocityX();
-            VelocityY = (int)(InitialVelocity * cosVal) + parentShip.GetVelocityY();
+            VelocityX = (InitialVelocity * sinVal) + parentShip.GetVelocityX();
+            VelocityY = (InitialVelocity * cosVal) + parentShip.GetVelocityY();
         }
 
         /// <summary>
